+++ R/geom_from_path.R
@@ -6,20 +6,20 @@
 #' @inheritParams ggplot2::geom_point
 #' @section Aesthetics:
 #' `geom_from_path()` understands the following aesthetics (required aesthetics are in bold):
-#' \itemize{
-#'   \item{**x**}{ - The x-coordinate.}
-#'   \item{**y**}{ - The y-coordinate.}
-#'   \item{**path**}{ - a file path, url, raster object or bitmap array. See [`magick::image_read()`] for further information.}
-#'   \item{`alpha = NULL`}{ - The alpha channel, i.e. transparency level, as a numerical value between 0 and 1.}
-#'   \item{`colour = NULL`}{ - The image will be colorized with this colour. Use the special character `"b/w"` to set it to black and white. For more information on valid colour names in ggplot2 see <https://ggplot2.tidyverse.org/articles/ggplot2-specs.html?q=colour#colour-and-fill>}
-#'   \item{`angle = 0`}{ - The angle of the image as a numerical value between 0° and 360°.}
-#'   \item{`hjust = 0.5`}{ - The horizontal adjustment relative to the given x coordinate. Must be a numerical value between 0 and 1.}
-#'   \item{`vjust = 0.5`}{ - The vertical adjustment relative to the given y coordinate. Must be a numerical value between 0 and 1.}
-#'   \item{`width = 1.0`}{ - The desired width of the image in `npc` (Normalised Parent Coordinates).
+#' \describe{
+#'   \item{**`x`**}{The x-coordinate.}
+#'   \item{**`y`**}{The y-coordinate.}
+#'   \item{**`path`**}{a file path, url, raster object or bitmap array. See [`magick::image_read()`] for further information.}
+#'   \item{`alpha = NULL`}{The alpha channel, i.e. transparency level, as a numerical value between 0 and 1.}
+#'   \item{`colour = NULL`}{The image will be colorized with this colour. Use the special character `"b/w"` to set it to black and white. For more information on valid colour names in ggplot2 see <https://ggplot2.tidyverse.org/articles/ggplot2-specs.html?q=colour#colour-and-fill>}
+#'   \item{`angle = 0`}{The angle of the image as a numerical value between 0° and 360°.}
+#'   \item{`hjust = 0.5`}{The horizontal adjustment relative to the given x coordinate. Must be a numerical value between 0 and 1.}
+#'   \item{`vjust = 0.5`}{The vertical adjustment relative to the given y coordinate. Must be a numerical value between 0 and 1.}
+#'   \item{`width = 1.0`}{The desired width of the image in `npc` (Normalised Parent Coordinates).
 #'                           The default value is set to 1.0 which is *big* but it is necessary
 #'                           because all used values are computed relative to the default.
 #'                           A typical size is `width = 0.1` (see below examples).}
-#'   \item{`height = 1.0`}{ - The desired height of the image in `npc` (Normalised Parent Coordinates).
+#'   \item{`height = 1.0`}{The desired height of the image in `npc` (Normalised Parent Coordinates).
 #'                            The default value is set to 1.0 which is *big* but it is necessary
 #'                            because all used values are computed relative to the default.
 #'                            A typical size is `height = 0.1` (see below examples).}