+++ NEWS.md
@@ -1,6 +1,7 @@
 # ggpath (development version)
 
 * Invalid paths no longer result in an error. Instead, ggpath will warn the user and insert an empty grob.
+* Minor adjustment to the format of lists in the documentation to avoid check problems in an upcoming R release.
 
 # ggpath 1.0.1